outsource the same components
