+++ app/src/main/java/hu/tb/minichefy/presentation/screens/components/EmptyContent.kt
@@ -1,0 +1,60 @@
+package hu.tb.minichefy.presentation.screens.components
+
+import androidx.compose.animation.animateColor
+import androidx.compose.animation.core.RepeatMode
+import androidx.compose.animation.core.infiniteRepeatable
+import androidx.compose.animation.core.rememberInfiniteTransition
+import androidx.compose.animation.core.tween
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.size
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.dp
+import com.airbnb.lottie.compose.LottieAnimation
+import com.airbnb.lottie.compose.LottieCompositionSpec
+import com.airbnb.lottie.compose.LottieConstants
+import com.airbnb.lottie.compose.rememberLottieComposition
+import hu.tb.minichefy.R
+
+@Composable
+fun EmptyContent(
+    descriptionResource: Int
+) {
+    val infiniteTransition = rememberInfiniteTransition(label = "infinite transition")
+    val animatedColor by infiniteTransition.animateColor(
+        initialValue = MaterialTheme.colorScheme.primary,
+        targetValue = MaterialTheme.colorScheme.secondary,
+        animationSpec = infiniteRepeatable(tween(1000), RepeatMode.Reverse),
+        label = "color animation"
+    )
+
+    val composition by rememberLottieComposition(
+        LottieCompositionSpec.RawRes(
+            R.raw.book
+        )
+    )
+    Column(
+        modifier = Modifier.fillMaxSize(),
+        horizontalAlignment = Alignment.CenterHorizontally,
+        verticalArrangement = Arrangement.Center
+    ) {
+        LottieAnimation(
+            composition = composition,
+            modifier = Modifier
+                .size(120.dp),
+            iterations = LottieConstants.IterateForever
+        )
+        Text(
+            text = stringResource(id = descriptionResource),
+            style = MaterialTheme.typography.bodyLarge,
+            color = animatedColor
+        )
+    }
+}
